+++ react/components/ProductsPositionOffsetContext.tsx
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react'
+
+const ProductsPositionOffsetContext = createContext(0)
+
+export const ProductsPositionOffsetProvider =
+  ProductsPositionOffsetContext.Provider
+
+export function useProductsPositionOffset(): number {
+  return useContext(ProductsPositionOffsetContext)
+}
+
+export default ProductsPositionOffsetContext